Rename models/points-prelevement.js into models/point-prelevement.js
