+++ lib/demarches-simplifiees/consolidate.js
@@ -4,7 +4,7 @@ import {chain} from 'lodash-es'
 import * as Dossier from '../models/dossier.js'
 import * as Preleveur from '../models/preleveur.js'
 import {getTerritoires} from '../models/territoire.js'
-import {findPointPrelevementByIdPoint} from '../models/points-prelevement.js'
+import {findPointPrelevementByIdPoint} from '../models/point-prelevement.js'
 
 export async function consolidateDossiers() {
   const dossiers = await Dossier.getUnconsolidatedDossiers()
