+++ lib/models/exploitation.js
@@ -3,7 +3,7 @@ import mongo, {ObjectId} from '../util/mongo.js'
 import {validateChanges, validateCreation} from '../validation/exploitation-validation.js'
 import createHttpError from 'http-errors'
 
-import {getPointPrelevement} from './points-prelevement.js'
+import {getPointPrelevement} from './point-prelevement.js'
 import {getPreleveur} from './preleveur.js'
 import {getNextSeqId} from '../util/sequences.js'
