+++ lib/models/stats.js
@@ -5,7 +5,7 @@ import {
   getPointsPrelevementFromTerritoire,
   getPointsPrelevement,
   decoratePointPrelevement
-} from './points-prelevement.js'
+} from './point-prelevement.js'
 
 import mongo from '../util/mongo.js'
