+++ lib/routes.js
@@ -18,7 +18,7 @@ import {
   updatePointPrelevement,
   deletePointPrelevement,
   getPointsPrelevementFromTerritoire
-} from './models/points-prelevement.js'
+} from './models/point-prelevement.js'
 
 import {
   getPreleveur,
